+++ addOns/scripts/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Error when trying to run an unsupported script type through the Automation Framework.
 
 ## [45.1.0] - 2024-03-25
 ### Added